pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "linkedin-network-cleaner"
version = "0.1.0"
description = "Extract, score, and clean your LinkedIn network using AI"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "anthropic>=0.40.0",
    "typer>=0.12.0",
]

[project.scripts]
linkedin-cleaner = "linkedin_network_cleaner.cli.app:main"

[project.optional-dependencies]
dev = ["pytest>=8.0"]
